Added 'Default' virtual call mechanism which should work across all architectures, that generates conflict resolution stubs that package arguments into structs and pass a pointer to the struct. At this point architecture specific virtual call mechanisms can be added to improve performance (though this currently isn't very important). Also added clang to be run for optimisations, since it seems to do much better than locic right now (presumably it interacts more closely with LLVM's optimisation pipeline).

+++ src/Locic/CodeGen/VTable.hpp
@@ -14,11 +14,11 @@ namespace Locic {
 		const size_t VTABLE_SIZE = 20;
 		
 		/**
-		 * All method names are 'hashed' to a 32 bit
+		 * All method names are 'hashed' to a 64 bit
 		 * unsigned integer, for use in virtual method
 		 * tables.
 		 */
-		typedef uint32_t MethodHash;
+		typedef uint64_t MethodHash;
 		
 		MethodHash CreateMethodNameHash(const std::string& methodName);
